move enum in constants
