+++ packages/web-app/src/modules/demand-alerts-views/DemandAlertsPage/DemandAlertsSetUp/utils.ts
@@ -1,4 +1,4 @@
-import { SubscribeToDemandAlertStatus } from '../../DemandAlertsStore'
+import { SubscribeToDemandAlertStatus } from "../../constants";
 
 export const createNewSubscriptionErrorTexts: Partial<Record<SubscribeToDemandAlertStatus, string>> = {
   [SubscribeToDemandAlertStatus.Failure]: 'Something went wrong. Please try again later',
