+++ packages/web-app/src/modules/demand-alerts-views/constants.ts
@@ -1,5 +1,4 @@
 import type { DemandTier } from '../demand-monitor-views/DemandMonitorStore'
-import { SubscribeToDemandAlertErrorType, SubscribeToDemandAlertStatus } from './DemandAlertsStore'
 
 export const demandSubscriptionsPath = '/api/v2/demand-monitor/subscriptions'
 
@@ -12,6 +11,27 @@ export const demandScenario: Record<DemandTier, DemandScenarioDropdownOption> = 
   high: { label: 'High Demand', value: 'high' },
   mid: { label: 'Moderate Demand', value: 'mid' },
   low: { label: 'Low Demand', value: 'low' },
+}
+
+export enum SubscribeToDemandAlertStatus {
+  Unknown = 'UNKNOWN',
+  Submitting = 'SUBMITTING',
+  Success = 'SUCCESS',
+  Failure = 'FAILURE',
+  AlreadyExists = 'FAILURE:SUBSCRIPTION-ALREADY-EXISTS',
+  InvalidGpu = 'FAILURE:INVALID-GPU',
+}
+
+export enum UnsubscribeFromDemandAlertStatus {
+  Unknown = 'UNKNOWN',
+  Submitting = 'SUBMITTING',
+  Success = 'SUCCESS',
+  Failure = 'FAILURE',
+}
+
+export enum SubscribeToDemandAlertErrorType {
+  AlreadyExists = 'subscription:create:already-exists',
+  InvalidGpu = 'subscription:create:invalid-gpu',
 }
 
 export const subscribeToDemandAlertStatusByErrorType: Record<
